backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
pyairtable
bcrypt
pydantic
pydantic[email]
pyjwt